+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Version 1.36 (not yet released)
 
-* Data Processor, inconsistent use of "e-mail" and "email", use the CiviCRM standard "email" by !79
+* Inconsistent use of "e-mail" and "email", use the CiviCRM standard "email" by !79
+* Default delimiter for CSV export is semi-colon, not comma by !81
 
 # Version 1.35