+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Version 1.37 (not yet released)
+
+* Ability to return contact as array in relashionship type
+
 # Version 1.36 (not yet released)
 
 * Inconsistent use of "e-mail" and "email", use the CiviCRM standard "email" by !79